Instance of ArrowChoice for SF.
This commit introduces an instance of arrow-choice for signal
functions. ArrowChoice allows us to use a signal function or other
depending on the value of a signal. For instance, one can now
write expressions like:

someDef = proc i -> do
  if (i > 5)
   then do
     x <- sf1 <- i
     returnA -< x
   else do
     x <- sf2 <- i
     returnA -< x

The point being that the choice of which sf to use, sf1 or sf2, is made at each
point based on the value of the input signal i.

There is an important design choice in this implementation, namely whether the
sf not chosen advances in time or not. For example, if the first time the time
delta is d1 and sf1 is chosen, then it will be "advanced" by d1 time and
evaluated in the future. If the next time the time delta is d2 and sf2 is
chosen, we have a choice of whether to move sf2 forward by d2 time only, or by
(d1 + d2).

The first alternative, implemented by this commit, means that the two sfs, sf1
and sf2, are not in sync, and they refer to different points in time. If this
introduces any potential problems, please open a bug report and try to detail
why it is a problem.