Application of an SF to a collection.
This commit exposes a new SF combinator parC :: SF a b -> SF [a] [b],
which applies an SF to every element of a collection. It is easier to
work with than parallel switching combinators, which have long and
undecypherable signatures.

It works under the assumption that the input list will not increase in size
over time (new inputs are dropped). Maybe a more robust one could be provided
which will deal with that case.