Introduces dependency on DeepSeq (see #5).
Yampa used to include its own version of 'deepseq', called 'Forceable'.
While code can use Forceable, it is better to rely on standard
implementations. That is why Forceable was deprecated in favour of
NFData, which defines an interchangeable interface. Forceable is
still included, but its use is discouraged and the compiler shows
a warning.

This commit includes a dependency on the deepseq package. The only
instance of Forceable that was included in Yampa was for Event,
which now instances NFData as well. Due to the change in API,
the version minor is increased (0.10.2).