Time dilation/compression.
It is sometimes useful to define a time transformation on a signal function,
for instance, to run it slower or faster than normally.

This commit introduces a new combinator that takes a time transformation
function and runs an SF affecting time at every step.

This is a very simple transformation: the transformation cannot depend on time,
on the present or past values, and it cannot reverse time (go backwards).