Tell travis to check only master branch.
We are experimenting with many different Yampa extensions.
As bad as it looks, sometimes we'll push ideas that are not
fully developed or depend on certain compiler versions or
a certain base.

To move more quickly, we need to be able to synchronise and
progress even without curating and maintaining our research
branches with the same care as the master branch.

If anyone wants to contribute to any of these branches, we can
adapt to that person's requirements by making sure that all
pushed commits compile also on their Haskell setup, and even
re-enable travis building for that particular branch.