ListSFs, aka branching SFs.
ListSFs are SFs that can produce multiple continuations. In Yampa, they are
implemented as SFs that produce an output, indicate whether they have finished,
and produce new ListSFs.

The idea is that a ListSF can branch off, and it can also die. If it branches
and later dies, its past children remain alive.

In games, this is useful to make a ball split in two, or a player produce a
bullet. In one case, the player produces a fire-event that needs to be
transformed by the caller SF into a new SF. With this approach, the player
itself can produce a bulletSF.

Details can be found at:

http://www.cs.nott.ac.uk/~psznhn/Talks/PHM-April2016-DeclarativeAbstractionsForGames.pdf

The game pang-a-lambda, also known as lambd-a-lot, uses this facility.

In the FRP implementation Dunai, a similar effect is achieved by using the list
monad.