Exposes SF and SF' constructors.
Some new extensions need to access SF and SF's directly, namely testing.

This patch exposes all the constructors as a facility *only*. It should never be integrated into the main tree.

In the future, it might be possible to control this using a flag in the cabal file. Unfortunately, there is an ongoing debate about whether it is a good idea to introduce flags that change a library's API.