Introduces testing features in separate library.
This commit introduces a separate library, currently in examples, that extends
Yampa with testing features. Yampa's Quickcheck regression test suite, which
was in tests/ before, is now part of that library as well.
It also minimizes the impact on Yampa itself, only requiring a few changes in
Simulation. There are different ways of implementing the one-step evaluation,
though, and I'm still undecided.