Include examples when compiling with flag. Refs #137.
The cabal file includes an examples flag that is supposed to compile the
examples in the examples dir, but not all were being compiled. In
particular, there are two examples, a tailgating detector and an
elevator example, of more historic significance due to their relation to
papers that used those examples to present key FRP ideas.  These two
examples were not being compiled and eventually broke when functions and
modules were reorganized, causing problems to some users, as detailed in
issue #137.

This patch adds those two examples to the cabal file as optional
executables. Because travis is already compiling everything with the
examples flag, they will also be checked every time.