Remove to modules from the haddock test. Refs #72.
Haddock has a problem with generating documentation for instances
and re-exported definitions.

Because our tests use haddock to check that all definitions are properly
documented, and because we use instances and re-export definitions from
System.Random and Data.VectorSpace, haddock reports a failure that makes
our test fail. This makes it impossible to add that test to travis to
make sure the whole library remains thoroughly documented.

This change performs three adaptations:

- Remove old modules that were excluded from the haddock check, because
they have been moved to a different library.

- Add FRP.Yampa and FRP.Yampa.Random to the exclusion list, because,
even through documented, haddock is not able to tell.

- Modify FRP.Yampa.Task to import definitions from the FRP.Yampa.*
landscape and not import FRP.Yampa, which is no longer available during
the haddock test due to being excluded.