Make passing the coverage tests a requirement. Refs #72.
The travis script currently considers haddock testing after_script only,
meaning that a failure in the haddock tests would not make the whole
build fail.

Because we have now reached a point in which haddock testing is
complete, we are ready to make documentation mandatory.

This change modifies the travis script so that it tests with haddock as
well on GHC 8.4.4 as part of the normal build steps, making full
documentation a requirement.