Added notice about implementing vector space using linear. Refs #132.
Yampa uses the library simple-affine-spaces as an abstraction over
vector spaces. The library provides the necessary type-classes, but
users can still implement their own instances.

Because other vector libraries are much more common than
simple-affine-spaces, it is helpful to make users aware that using Yampa
does not mean having to stick to that library, and they can trivially
define the necessary instances to use Yampa constructs like integral
with their preferred vector implementations.

This commit adds a notice to the Yampa module and to the integration
module, where users are most likely to benefit from that knowledge.