Add cabal.project to be used with new-build command. Refs #99.
Newer versions of cabal support having a cabal.project file listing
all packages and sources in a project, and to install all of them
with just one line.

This commit adds a cabal.project to the root of the tree, which adds
Yampa and yampa-test. It might be a better structure to have a yampa/
and a yampa-test/ folder.