yampa: Introduce benchmark. Refs #167.
This commit implements a basic benchmark that is integrated in the Yampa
package. The goal is to have an initial benchmark infrastructure that
can be used to evaluate the performance of Yampa, and compare between
different versions, and to make evidence-based decisions about when new
features should be incorporated into Yampa.