Update gitignore wrt. the gitignore project. Refs #158
Problem

The file
https://raw.githubusercontent.com/github/gitignore/master/Haskell.gitignore
contains many more entries than the .gitignore file in the top dir.

This has recently started being a problem because new versions of cabal
(?) generate additional files not included in that .gitignore.

Solution

Updated the .gitignore file to merge the records taken from the
gitignore project (under Taken from the gitignore project) with the
preexisting ones (under Others)

EDIT (iperez): added the description from the PR on github to the commit
message.