Fix broken links in README. Refs #178.
In #176 the structure of the repo changed. The README contained a link
to the yampa-test library that was left behind and now leads nowhere.
The same happened in dunai, which invalidated one of the links in this
README.

This commit updates those links to point directly to the corresponding
path in the repo on github.

[ci skip]