Declare modules in test component in Cabal file. Refs #181.
Some of the test modules are needed for compilation of the tests but not
listed in the `other-modules:` field. Cabal reports a warning about this
during compilation.

This commit modifies the Cabal file to list all modules used by the
tests in the appropriate components and fields. It has been tested with
cabal-install 2.4.