Adapt installation instructions to new cabal. Refs #186.
The README mentions that the library can be installed with cabal install
Yampa. While true, this will not give users of current versions of cabal
access to it in their GHCi installations.

This commit adapts the installation instructions to use --lib when
installing Yampa without the demos. That makes the library accessible in
the default environment loaded by GHCi, letting inexperienced users run
examples directly from the terminal.

[ci skip]