Add screenshots of games in production, add "new" game. Refs #188.
The README, in 'Use in production', only talks about Magic Cookies, but
another game was released in 2020.

This commit adds Enpuzzled to the list of Yampa related projects, and
modifies the use in production section to have screencaps like with
other games, and links to their websites and download sites.

[ci skip]