Correct module/copyright info in haddock documentation. Refs #192.
Some modules have the module, copyright, etc info before the module's
documentation's haddock marker (i.e., -- |), which makes that
information not be parsed or used by haddock.

This commit corrects the haddock documentation for all modules that had
no module/copyright/etc info in the documentation (either because it was
using the wrong haddock syntax, because it was missing, or because it
was separate from the module haddock comment).