In tests, Rename AFRP -> Yampa or remove. Refs #195.
The battery of regression tests contains the word AFRP everywhere, which
was the original name of Yampa. To bring all code up to date, that term
needs to be renamed or, where it is not needed, just removed.

This commit renames all tests modules, adapts all imports, and
edits the word AFRP or A F R P in the code itself (including
comments).

The Cabal file is adapted accordingly, and tested.