Remove incorrect test. Refs #198.
One of the quickcheck tests compares the result of applying two
exponentiation functions on floating-point numbers. The two
functions in question are (**) and (^).

While in principle these should be equivalent, they are not so
in practice. For example, for floating-point numbers, the following
two cases fail:
-13.4051075, with 179.69691 (**) and 179.6969 (^)
15.857409, with 251.45741 (**) and 251.4574 (^)

Running a test on all floats:

```haskell
import Numeric.IEEE
import Control.Monad
import System.IO
import Data.IORef

main = do
  ref <- newIORef (0 :: Int)
  forM_ list $ \l -> do
    let x1 = l**2
        x2 = l^2
        diff = x1 /= x2

    counter <- readIORef ref
    if (counter == 10000)
      then do
        -- print progress every few steps
        print (l, x1, x2)
        writeIORef ref 0
      else
        writeIORef ref (counter + 1)

    when diff $ hPutStrLn stderr (show (l, x1, x2))

list :: [Float]
list = starting : prevList starting
  where
    starting = sqrt maxFinite

prevList :: Float -> [Float]
prevList n
    | identicalIEEE n n' = []
    | otherwise          = n' : prevList n'
  where
    n' = predIEEE n
```

shows that there are over 774000 positive Floats for which the two
squaring implementations fail. The highest positive number for which
the program above finds a difference 1.8429595e19, and
the difference for that number is in the order of e31.

Running a similar test for Double numbers shows that, for example, for
the number 1.3407807830046642e154, there is a difference of
1.99584030953472e292.

All of the above makes the idea of implementing this test using any
notion of equality, or "approximate equality", simply not viable.

This commit removes such a test, since the arrow laws are covered by
other quickcheck tests and this was not part of Yampa's original
regression tests.