Remove unused language extensions. Refs #199.
Some modules declare GHC extensions that are not really needed in the
code inside the module. This is likely a vestige from when all code
lived in one main FRP module.

Removing unnecessary extensions will make the code easier to maintain,
read, and also help remove the use of unnecessary extensions in the
future if possible (which would make the code more portable).

This commit removes unused GHC extensions from all Yampa modules.