Move regression test suite with to yampa-test. Refs #201.
As Yampa progresses, as much of the regression tests as possible should
use the same mechanisms for testing.

Yampa's regression tests are currently in yampa/, while some randomized
tests are with yampa-test. The randomized test cover features also
tested in the regression tests, so the latter should progressively
disappear in favor of randomized tests when coverage is deemed similar
or better with the new method.

This commit moves all tests into yampa-test, to progress in a direction
of uniformity, test coverage, and simplification.