yampa-test: Disable flag by default. Refs #208.
The space usage tests require users to monitor the space consumed by the
application as they run the test.

If therefore makes no sense for that test to be ran automatically, since
most users will either be just installing Yampa, or just checking that
Yampa's tests finish successfully.

This commit makes the default status of that flag false, requiring that
users enable it manually if they really want to run the space tests.