Remove vim customizations from Haskell code. Refs #209.
Yampa files contain comments with Vim commands so that tabs are
expanded.

Tabs are currently not allowed, and will be flagged as a warning. In the
future, we will likely make that an error.

At this stage in software development, using tabs is the oddity.
Instead, users should ensure that they customize their own vim script
file to expand tabs in all files, or otherwise ensure they do not
introduce them when contributing to other people's projects.

This commit removes all settings introduced at the bottom of files to
indicate to vim that tabs should be expanded.