Rename module to reflect that it now focuses on space leaks. Refs #216.
In a prior commit, all tests pertaining to functional correctness were
moved out of a module and into a dedicated module matching Yampa's
current module hierarchy. As a consequence, all tests that remain in the
old module now pertain to space leaks.

This commit renames the module to better match its narrower purpose.