Move tests into modules matching Yampa's module hierarchy (loopIntegral). Refs #216.
Tests are grouped by Yampa's old module hierarchy, making it difficult
to understand which portions of Yampa are tested, and which ones are
not. Also, the YampaQC.hs module includes tests that are also covered,
in total or in part, in other regression testing modules.

This commit moves tests to better match Yampa's current module
hierarchy.