yampa-test: Organize module Switches following Yampa's order. Refs #217.
Understanding which parts of Yampa are tested and which ones are not is
generally hard. To facilitate keeping track of which elements are tested
and which ones are not, this commit organizes the elements of Switches
in the same order and with the same haddock section separators as they
have in Yampa's counterpart.