Move ArrowLoop tests into InternalCore test module. Refs #218.
The unit tests for the ArrowLoop interface were in issue #217 placed in
Test.FRP.Yampa.Loop. However, the ArrowLoop instance is defined in
FRP.Yampa.InternalCore, so those tests must be moved.

This commit moves all tests pertaining to the ArrowLoop interface to the
end of InternalCore, matching the order in which the instance is defined
in the corresponding Yampa module.