yampa: Make use of arrows less explicit in documentation. Refs #183.
Although Yampa provides arrow instances and is very often used with
arrows, they are not really necessary. One can also use the applicative
interface, Applicative Do and (category theoretic) composition, plus all
the combinators included with the library.

To facilitate adoption, this commit modifies the description in the
cabal file, main haskell file, and README, to make the use of Arrows
less prominent and indicate to users that using other styles is also
supported.