Replace all occurrences of 'AFRP' with 'Yampa'. Refs #223.
Error messages produced by Yampa print "AFRP", the original name of the
library, as opposed to "Yampa". This is likely confusing for new users,
since there connection between those two is not explained in the README
or in the documentation of the library.

This commit replaces all occurrences of AFRP/AFRPTask with
Yampa/YampaTask.