yampa-test: Complete Tests for FRP.Yampa.Basic. Refs #219.
This commit completes the unit tests for the functions in
FRP.Yampa.Basic without a unit test.

It also re-implements the unit tests for (-->) to use integral, to
distinguish it from other functions in the same module that exhibit the
same point-wise (or timewise) behavior but different temporal behavior.

The patch also adds the subsection headers to functions in the module,
so that each function tested has its own subsection header/separator.