yampa-test: Complete Tests for FRP.Yampa.Conditional. Refs #225.
This commit completes the unit tests for the functions in
FRP.Yampa.Conditional without a unit test.

The implementation of the test for pause is based on a model function of
pause that uses ArrowLoop and ArrowCase. It is actually quite tricky to
get right, which suggests that 1) we may need a simpler way to visualize
and apply signal functions, or 2) perhaps pause is just too complex.