yampa-test: Complete tests for FRP.Yampa.Delays. Refs #226.
This commit completes the unit tests for the functions in
FRP.Yampa.Delays without a unit test.

The implementation of the tests for pre and fby are based on a model
function of iPre. It is tricky to implement a test for the first value
of pre, since it is supposed to be uninitialized (and we can't test for
that, not easily).