yampa-test: Adjust years, copyright owners. Refs #235.
This project borrows from the Yampa project with the inclusion of the
tests as part of the library.

This commit adjusts the copyright holders listed in the haddock comments
of the tests to match the main developers that contributed to those
tests over those years.

The license file has been modified to also mention the fact that this
project borrows from Yampa.