yampa: Add version bounds to dependencies. Refs #233.
The cabal file in Yampa does not specify upper bounds on dependencies.
This is bad practice and should be considered a bug. It is currently
limiting or endangering an update to a dependency, which would likely
break packages created by Yampa users.

This commit introduces version bounds for all dependencies of Yampa, all
tests, and all executables/demos.