yampa: Replace broken links. Refs #253.
There are multiple references to the Yampa Arcade paper pointing to
Yale's Haskell server in the CS department. That server is down so the
links no longer work.

This commit replaces all links to the same paper on Antony Courtney's
website, which is currently working. We specifically do not point users
to the ACM website directly because the paper is not freely accessible
there as of the time of this writing.