yampa: Format comment as haddock comment (KSG Haskell 1.3.0 - 3.3). Refs #255.
All Haskell files must include a copyright notice in a comment. The
copyright notice must be included in a format parseable by Haddock.

Haddock does not have an Authors field for module declarations, but the
original file did. To retain that information in some way, we decide to
make this file Copyright of Henrik Nilsson only upon creation.