yampa-test: List testing example in cabal file. Refs #257.
The Testing file containing examples from the ICFP 2017 FRP testing
paper is not listed in the cabal file. This makes the example not be
included with the source distribution file, and not uploaded to hackage.

This commit lists the example in the cabal file so that it is included
in the source distribution file for the package.