yampa-test: Remove redundant pragma (KSG Haskell 1.3.0 - 3.6.4). Refs #256.
No redundant (unused) definitions should be included. A definition may be
redundant because it is internal to a module or function and not used in
the module, or because it is not used by any project that uses a library
or module.

This commit eliminates an unnecessary MultiwayIf, by simplifying the
if-then-else that used it.