yampa: Re-format CHANGELOG to remove paths. Refs #261.
The CHANGELOGs list the path associated to a change, prior to the
summary of the change. The addition of paths to those changes adds very
little useful information (if any at all). The decision of how big a
part of the path to show in the CHANGELOG (e.g., src/FRP/Yampa.hs vs
src/) is quite arbitrary. Adding the paths also implies that, sometimes,
a change may have to be listed multiple times, since it affected
multiple paths, and the re-wrapping of change summaries to fit in one
line makes for awkward diffs and makes reading the CHANGELOGs harder.

This commit removes the paths from the entries in Yampa's CHANGELOG, and
reformats the entries to be one per line. Because the removal of the
path also removes some context from each entry, some entries are
adjusted to better reflect the change. Whenever needed, the summary from
the accompanying issue on github is used to re-phrase the entry.