yampa: Remove vertical space in CHANGELOG. Refs #261.
The CHANGELOGs current list each release in a separate group, prefaced
by the maintainer's details and release date. Sometimes, a space is left
after the maintainer's details and sometimes no space is left. To
facilitate reading the groups, we want to stick to a standard format as
much as possible.

This commit removes that space for each group, which makes it clearer
that the group of entries follow the preceding release date line.