yampa-test: Re-format CHANGELOG to remove paths. Refs #261.
The CHANGELOGs list the path associated to a change, prior to the
summary of the change. The addition of paths to those changes adds very
little useful information (if any at all). The decision of how big a
part of the path to show in the CHANGELOG (e.g., src/FRP/Yampa.hs vs
src/) is quite arbitrary. Adding the paths also implies that, sometimes,
a change may have to be listed multiple times, since it affected
multiple paths, and the re-wrapping of change summaries to fit in one
line makes for awkward diffs and makes reading the CHANGELOGs harder.

This commit removes the paths from the entries in yampa-test's
CHANGELOG, and reformats the entries to be one per line.