yampa: Refine description of new contribution process in README. Refs #265.
The Contributions section of the README directs users to open new issues
with a number of possible topics. The way it is currently phrased, it
seems like open an new issue is the may way to interact with the
project.

This commit updates the readme instructing users to use discussions by
default, and only issues when there's a specific bug to fix or new
feature to include.