yampa: Clarify what to include in PR in README. Refs #265.
The contribution process (specifically, commit/PR handling) does not
specify that the change to the CHANGELOG should be separate, and that
packages should be named in the summary line. This would likely be one
of the most likely reasons why a PR would need to be amended or
rejected.

This commit modifies the contribution process description in the README
to make these two additional requirements explicit.