yampa: Define Yampa.FRP.Task.return in terms of pure. Refs #276.
The monad instance for Yampa.FRP.Task.Task does not define return in
terms of the Applicative pure. This is leading to a warning due to
upcoming changes in GHC. The definition is the same as that of pure, so
there's no reason not to rely on it.

This commit modifies the Monad instance of Yampa.FRP.Task.Task to define
return in terms of the Applicative pure.