yampa: Make FRP.Yampa.Event.maybeToEvent public. Refs #267.
The function FRP.Yampa.Event.maybeToEvent is labelled as internal, yet
it is being exported by the module. The function is useful and is being
used in the wild, so removing it now would break existing applications.

This commit makes the function "officially" public. It changes the
comment stating that it is internal, and it moves it within the module
to a location that makes more sense in terms of explaining the API.