yampa: Simplify FRP.Yampa.Switches.safeZip. Refs #298.
The function FRP.Yampa.Switches.safeZip, or more specifically, safeZip'
inside it, captures a list and then applies some head and tail
alternative functions to it.

This commit reduces the function by pattern matching on the list,
already obtaining the head and the tail. We move the error reporting to
a third, default case.