yampa: Implement integral using trapezoid rule. Refs #263.
The current implementation of integral has a tendency to diverge. This
commonly manifests in games as bouncing elements that bounce more and
more over time, or springs that bounce harder. In the general case, the
energy of the simulated system increases over time.

This commit introduces a new integral function that uses the trapezoid
rule. The new integral can be more accurate and does not increase over
time in the same way that the rectangle rule does.

Co-authored-by: Dominic Steinitz <dominic@steinitz.org>
Co-authored-by: Miguel Negrao <miguel.negrao@friendlyvirus.org>